Update index.ts
Adds a value of Submit to the batch job submit buttons. Default appearance on Firefox (68.0.1) is "Submit Query", which is confusing for a user. This PR adds an explicit value of "Submit", so behaviour on Firefox is aligned with the default behaviour on Safari (12.1.2) and Chrome (76.0.3809.132) on Mac OSX, which was "Submit" anyway.